refactor(react-google-adk): move the pure adk helpers to the converter seam
