+++ .changeset/adk-converter-seam.md
@@ -1,0 +1,5 @@
+---
+"@assistant-ui/react-google-adk": patch
+---
+
+refactor: move the pure adk helpers to the converter seam.
